+++ src/main/java/aleetcode/baidu/IntegerTest.java
@@ -1,0 +1,12 @@
+package aleetcode.baidu;
+
+public class IntegerTest {
+    public static void main(String[] args) {
+        Integer a = -128;
+        Integer b = -128;
+        System.out.println(a == b);
+        a = 127;
+        b = 127;
+        System.out.println(a == b);
+    }
+}
