+++ src/main/java/aleetcode/contest2020/Contest197.java
@@ -1,4 +1,4 @@
-package aleetcode;
+package aleetcode.contest2020;
 
 import java.util.*;
