+++ src/main/java/aleetcode/util/TreeNode.java
@@ -1,11 +1,20 @@
 package aleetcode.util;
 
 public class TreeNode {
-	public int val;
-	public TreeNode left;
-	public TreeNode right;
+    public int val;
+    public TreeNode left;
+    public TreeNode right;
 
-	public TreeNode(int x) {
-		val = x;
-	}
+    public TreeNode() {
+    }
+
+    public TreeNode(int x) {
+        val = x;
+    }
+
+    public TreeNode(int val, TreeNode left, TreeNode right) {
+        this.val = val;
+        this.left = left;
+        this.right = right;
+    }
 }
