+++ src/main/java/thinkinjava/reflect/DynamicProxyHandler.java
@@ -1,0 +1,21 @@
+package thinkinjava.reflect;
+
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+
+public class DynamicProxyHandler implements InvocationHandler {
+
+    private Object proxied;
+
+    public DynamicProxyHandler(Object proxied) {
+        this.proxied = proxied;
+    }
+
+    @Override
+    public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+//        System.out.println("proxy = " + proxy + ", method = " + method + ", args = " + Arrays.deepToString(args));
+        return method.invoke(proxied, args);
+    }
+
+}
