+++ src/main/java/thinkinjava/reflect/Interface.java
@@ -1,0 +1,7 @@
+package thinkinjava.reflect;
+
+public interface Interface {
+    void doSomething();
+
+    void somethingElse(String arg);
+}
