+++ src/main/java/thinkinjava/reflect/RealObject.java
@@ -1,0 +1,14 @@
+package thinkinjava.reflect;
+
+public class RealObject implements Interface {
+
+    @Override
+    public void doSomething() {
+        System.out.println("doSomething");
+    }
+
+    @Override
+    public void somethingElse(String arg) {
+        System.out.println("somethingElse" + arg);
+    }
+}
