+++ src/main/java/work/redis/RedisLock.java
@@ -1,0 +1,71 @@
+package work.redis;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import redis.clients.jedis.Jedis;
+import redis.clients.jedis.JedisPool;
+import redis.clients.jedis.params.SetParams;
+
+import java.util.Collections;
+
+//@Service
+public class RedisLock {
+
+    Logger logger = LoggerFactory.getLogger(this.getClass());
+
+    private String lock_key = "redis_lock"; //锁键
+
+    protected long internalLockLeaseTime = 30000;//锁过期时间
+
+    private long timeout = 999999; //获取锁的超时时间
+
+
+    //SET命令的参数 
+    SetParams params = SetParams.setParams().nx().px(internalLockLeaseTime);
+
+    @Autowired
+    JedisPool jedisPool;
+
+
+    /**
+     * 加锁
+     *
+     * @param id
+     * @return
+     */
+    public boolean lock(String id) {
+        try (Jedis jedis = jedisPool.getResource()) {
+            String lock = jedis.set(lock_key, id, params);
+            if ("OK".equals(lock)) {
+                return true;
+            }
+            return false;
+        }
+    }
+
+    /**
+     * 解锁
+     * @param id
+     * @return
+     */
+    public boolean unlock(String id){
+        Jedis jedis = jedisPool.getResource();
+        String script =
+                "if redis.call('get',KEYS[1]) == ARGV[1] then" +
+                        "   return redis.call('del',KEYS[1]) " +
+                        "else" +
+                        "   return 0 " +
+                        "end";
+        try {
+            Object result = jedis.eval(script, Collections.singletonList(lock_key),
+                    Collections.singletonList(id));
+            if("1".equals(result.toString())){
+                return true;
+            }
+            return false;
+        }finally {
+            jedis.close();
+        }
+    }
+}
